requirements.txt
django
south
pytz
ipython
requests
TwitterAPI
python-dateutil
